# backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Utang API
  description: >
    MVP API for tracking customer utang (debt), recording payments, generating
    copyable reminders. Ledger-first design.
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Auth
  - name: Customers
  - name: Ledger
  - name: Reminders
  - name: Public

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a store owner (username + password + mobile)
      description: >
        Creates the store account and signs the owner in on the current device.
        Send an `X-Device-Id` header to trust this device (skips OTP on future logins).
        Store name, username and mobile number must each be unique; a conflict
        returns 409 with a deliberately generic message that does not reveal
        which detail is already in use.
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered and authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          $ref: '#/components/responses/Error'

  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with username + password
      description: >
        On a trusted device (matching `X-Device-Id`) returns status `AUTHENTICATED`
        with a token. From a new device returns status `OTP_REQUIRED` and sends an
        OTP to the owner's mobile; complete via `/auth/verify-device`.
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated, or an OTP challenge is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Error'

  /auth/verify-device:
    post:
      tags: [Auth]
      summary: Verify a new-device OTP and obtain a session token
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyDeviceRequest'
      responses:
        '200':
          description: Authenticated; device is now trusted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Error'

  /me:
    get:
      tags: [Auth]
      summary: Current store
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponse'

  /store:
    put:
      tags: [Auth]
      summary: Update store details
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoreRequest'
      responses:
        '200':
          description: Updated store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponse'
        '401':
          $ref: '#/components/responses/Error'

  /store/phone/verify/request:
    post:
      tags: [Auth]
      summary: Send an OTP to verify ownership of the store's mobile number
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OTP sent to the store's mobile number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneVerificationResponse'
        '401':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'

  /store/phone/verify/confirm:
    post:
      tags: [Auth]
      summary: Confirm the OTP and mark the mobile number as verified
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneVerificationRequest'
      responses:
        '200':
          description: Phone number verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponse'
        '401':
          $ref: '#/components/responses/Error'

  /store/qr:
    get:
      tags: [Auth]
      summary: Get the store's payment QR code image
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The QR code image
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: No QR code uploaded
    put:
      tags: [Auth]
      summary: Upload or replace the store's payment QR code image
      description: >
        Accepts a PNG, JPG, WebP or GIF image up to 2 MB. Replaces any existing
        QR code. The image is stored inline with the store record.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: QR code saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
    delete:
      tags: [Auth]
      summary: Remove the store's payment QR code image
      security:
        - bearerAuth: []
      responses:
        '200':
          description: QR code removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponse'
        '401':
          $ref: '#/components/responses/Error'

  /customers:
    post:
      tags: [Customers]
      summary: Add a customer (suki)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '200':
          description: Created customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
    get:
      tags: [Customers]
      summary: List customers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'

  /customers/{id}:
    get:
      tags: [Customers]
      summary: Get a customer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          $ref: '#/components/responses/Error'
    put:
      tags: [Customers]
      summary: Update a customer's name and mobile number
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          $ref: '#/components/responses/Error'

  /ledger/debit:
    post:
      tags: [Ledger]
      summary: Add utang (DEBIT)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LedgerMutationRequest'
      responses:
        '200':
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'

  /ledger/credit:
    post:
      tags: [Ledger]
      summary: Record bayad (CREDIT) — e.g. cash payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LedgerMutationRequest'
      responses:
        '200':
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'

  /customers/{id}/ledger:
    get:
      tags: [Ledger]
      summary: Ledger history for a customer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Ledger
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerResponse'

  /customers/{id}/reminder-preview:
    get:
      tags: [Reminders]
      summary: Preview a reminder message
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderPreviewResponse'

  /customers/{id}/remind:
    post:
      tags: [Reminders]
      summary: Log a reminder (copy). Max 1 per customer per day.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Reminder logged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemindResponse'
        '409':
          $ref: '#/components/responses/Error'

  /public/pay/{token}:
    get:
      tags: [Public]
      summary: Public payment page data (no auth)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public pay info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicPayResponse'
        '404':
          $ref: '#/components/responses/Error'

  /public/pay/{token}/qr:
    get:
      tags: [Public]
      summary: Store's payment QR code image (no auth)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: QR code image
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: Store has no QR code

  /public/pay/{token}/notify-paid:
    post:
      tags: [Public]
      summary: Text the store owner that the customer has paid (no auth)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaidNotificationResponse'
        '429':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  parameters:
    CustomerId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    DeviceId:
      name: X-Device-Id
      in: header
      required: false
      description: Client-generated device identifier used to trust the device.
      schema:
        type: string

  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    RegisterRequest:
      type: object
      required: [username, password, phoneNumber, storeName]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 60
          pattern: "^[a-zA-Z0-9]+$"
          example: "alingnena"
        password:
          type: string
          format: password
          minLength: 6
          example: "s3cret!"
        phoneNumber:
          type: string
          example: "09171234567"
        storeName:
          type: string
          example: "Aling Nena Store"
        ownerName:
          type: string
          nullable: true
          example: "Nena Reyes"
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: "aling.nena"
        password:
          type: string
          format: password
    VerifyDeviceRequest:
      type: object
      required: [username, code]
      properties:
        username:
          type: string
        code:
          type: string
          example: "123456"
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        store:
          $ref: '#/components/schemas/StoreResponse'
    LoginResponse:
      type: object
      properties:
        status:
          type: string
          enum: [AUTHENTICATED, OTP_REQUIRED]
        token:
          type: string
          nullable: true
          description: Present when status is AUTHENTICATED
        store:
          $ref: '#/components/schemas/StoreResponse'
        phoneNumber:
          type: string
          nullable: true
          description: Masked mobile number the OTP was sent to (when OTP_REQUIRED)
        devCode:
          type: string
          nullable: true
          description: Echoed OTP when no SMS gateway is configured; null when sent via SMS
    UpdateStoreRequest:
      type: object
      required: [storeName, phoneNumber]
      properties:
        storeName:
          type: string
          example: "Aling Nena Store"
        ownerName:
          type: string
          nullable: true
          example: "Nena Reyes"
        phoneNumber:
          type: string
          description: Changing this resets phoneVerified to false until re-verified.
          example: "09171234567"
    PhoneVerificationRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          example: "123456"
    PhoneVerificationResponse:
      type: object
      properties:
        phoneNumber:
          type: string
          description: Masked target number the OTP was sent to.
          example: "+63•••••4567"
        devCode:
          type: string
          nullable: true
          description: Echoed only when SMS delivery is not configured.
    StoreResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
          nullable: true
        phoneNumber:
          type: string
        name:
          type: string
        ownerName:
          type: string
          nullable: true
        onboarded:
          type: boolean
        phoneVerified:
          type: boolean
        hasQrCode:
          type: boolean
          description: Whether a payment QR code image has been uploaded (fetch it from GET /store/qr).
    CreateCustomerRequest:
      type: object
      required: [name, phoneNumber]
      properties:
        name:
          type: string
        phoneNumber:
          type: string
    UpdateCustomerRequest:
      type: object
      required: [name, phoneNumber]
      properties:
        name:
          type: string
        phoneNumber:
          type: string
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        phoneNumber:
          type: string
          nullable: true
        currentBalance:
          type: number
        payToken:
          type: string
    LedgerMutationRequest:
      type: object
      required: [customerId, amount]
      properties:
        customerId:
          type: integer
          format: int64
        amount:
          type: number
        note:
          type: string
          nullable: true
    LedgerEntryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [DEBIT, CREDIT]
        amount:
          type: number
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    LedgerResponse:
      type: object
      properties:
        customerId:
          type: integer
          format: int64
        currentBalance:
          type: number
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LedgerEntryResponse'
    ReminderPreviewResponse:
      type: object
      properties:
        message:
          type: string
        canSendToday:
          type: boolean
    RemindResponse:
      type: object
      properties:
        message:
          type: string
        sent:
          type: boolean
    PublicPayResponse:
      type: object
      properties:
        storeName:
          type: string
        customerName:
          type: string
        outstandingBalance:
          type: number
        storeHasQrCode:
          type: boolean
        history:
          type: array
          items:
            $ref: '#/components/schemas/PublicLedgerEntry'
    PaidNotificationResponse:
      type: object
      properties:
        sent:
          type: boolean
    PublicLedgerEntry:
      type: object
      properties:
        type:
          type: string
          enum: [DEBIT, CREDIT]
        amount:
          type: number
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
        error:
          type: string
        message:
          type: string